Updated README.md to make it more user-friendly
The instructions were really unclear, especially to a newcomer. We as a whole should aim to make the open source community more open (pun intended) to the beginners who want to contribute and learn with us.
The installation manual was lacking the command "npm i" that would make sure all the required dependencies are installed and up-to-date, as well as a note informing the user that NPM-JS is needed.